[add] env.go for handling all environment variable related logic, update all other modules accordingly.
