+++ internal/shells/bash.go
@@ -1,53 +1,16 @@
 package shells
 
 import (
-	"bufio"
 	"bytes"
 	"fmt"
 	"os"
 	"os/exec"
-	"regexp"
 	"strings"
 
 	"golang.org/x/sys/unix"
 
 	"github.com/Azure/InnovationEngine/internal/lib"
-	"github.com/Azure/InnovationEngine/internal/lib/fs"
 )
-
-// Location where the environment state from commands is captured and sent to
-// for being able to share state across commands.
-var environmentStateFile = "/tmp/env-vars"
-
-func loadEnvFile(path string) (map[string]string, error) {
-	if !fs.FileExists(path) {
-		return nil, fmt.Errorf("env file '%s' does not exist", path)
-	}
-
-	file, err := os.Open(path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open env file '%s': %w", path, err)
-	}
-
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	env := make(map[string]string)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.Contains(line, "=") {
-			parts := strings.SplitN(line, "=", 2) // Split at the first "=" only
-			value := parts[1]
-			if len(value) >= 2 && value[0] == '"' && value[len(value)-1] == '"' {
-				// Remove leading and trailing quotes
-				value = value[1 : len(value)-1]
-			}
-			env[parts[0]] = value
-		}
-	}
-	return env, nil
-}
 
 func appendToBashHistory(command string, filePath string) error {
 	file, err := os.OpenFile(filePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
@@ -70,46 +33,6 @@ func appendToBashHistory(command string, filePath string) error {
 	}
 
 	return nil
-}
-
-// Resets the stored environment variables file.
-func ResetStoredEnvironmentVariables() error {
-	return os.Remove(environmentStateFile)
-}
-
-var environmentVariableName = regexp.MustCompile("^[a-zA-Z_][a-zA-Z0-9_]*$")
-
-func filterInvalidKeys(envMap map[string]string) map[string]string {
-	validEnvMap := make(map[string]string)
-	for key, value := range envMap {
-		if environmentVariableName.MatchString(key) {
-			validEnvMap[key] = value
-		}
-	}
-	return validEnvMap
-}
-
-func CleanEnvironmentStateFile() error {
-	env, err := loadEnvFile(environmentStateFile)
-	if err != nil {
-		return err
-	}
-
-	env = filterInvalidKeys(env)
-
-	file, err := os.Create(environmentStateFile)
-	if err != nil {
-		return err
-	}
-
-	writer := bufio.NewWriter(file)
-	for k, v := range env {
-		_, err := fmt.Fprintf(writer, "%s=\"%s\"\n", k, v)
-		if err != nil {
-			return err
-		}
-	}
-	return writer.Flush()
 }
 
 type CommandOutput struct {
@@ -135,7 +58,7 @@ func executeBashCommandImpl(
 		"set -e",
 		command,
 		"IE_LAST_COMMAND_EXIT_CODE=\"$?\"",
-		"env > " + environmentStateFile,
+		"env > " + lib.DefaultEnvironmentStateFile,
 		"exit $IE_LAST_COMMAND_EXIT_CODE",
 	}
 
@@ -164,7 +87,7 @@ func executeBashCommandImpl(
 	// after a command is executed within a file and then loading that file
 	// before executing the next command. This allows us to share state between
 	// isolated command calls.
-	envFromPreviousStep, err := loadEnvFile(environmentStateFile)
+	envFromPreviousStep, err := lib.LoadEnvironmentStateFile(lib.DefaultEnvironmentStateFile)
 	if err == nil {
 		merged := lib.MergeMaps(config.EnvironmentVariables, envFromPreviousStep)
 		for k, v := range merged {
